Changing time window
Previously we were looking at only the most recent history. This will most often return "insufficient history" which isn't a success or failure. It just means it doesn't know enough to make a report for today. 

This is bad because something could be failing for days and fail to report as failed because it will always have a most recent status of insufficient history. 

Therefore, we are changing the timeWindow from Between0dAnd1d to Between1dAnd2d